+++ imod/tests/test_mf6/test_mf6_regrid_model.py
@@ -4,7 +4,6 @@ import numpy as np
 import pytest
 
 import imod
-from imod.mf6.lak import Lake
 from imod.tests.fixtures.mf6_small_models_fixture import (
     grid_data_structured,
     grid_data_unstructured,
@@ -75,22 +74,6 @@ def test_regrid_unstructured_model_to_unstructured_model(
     assert len(new_gwf_model.items()) == len(unstructured_flow_model.items())
     validation_result = new_gwf_model.validate("test_model")
     assert not validation_result.has_errors()
-
-
-def test_regrid_model_with_unsupported_package(
-    unstructured_flow_model, naardermeer, ijsselmeer
-):
-    lake_pkg = Lake.from_lakes_and_outlets(
-        [naardermeer(), ijsselmeer()],
-    )
-    unstructured_flow_model["lake"] = lake_pkg
-
-    finer_idomain = grid_data_unstructured(np.int32, 1, 0.4)
-    with pytest.raises(
-        NotImplementedError,
-        match="regridding is not implemented for package lake of type <class 'imod.mf6.lak.Lake'>",
-    ):
-        _ = unstructured_flow_model.regrid_like(finer_idomain)
 
 
 @pytest.mark.parametrize("inactivity_marker", [0, -1])