+++ docs/api/changelog.rst
@@ -52,6 +52,9 @@ Added
 - added logging to various initialization methods, write methods and dump methods. `See the
   documentation <https://deltares.github.io/imod-python/api/generated/logging/imod.logging.html>`_
   how to activate logging.
+- added :func:`imod.data.hondsrug_simulation` and
+  :func:`imod.data.hondsrug_crosssection` data.
+- simulations and models that include a lake package now raise an exception on clipping, partitioning or regridding. 
 
 Changed
 ~~~~~~~